+++ include/ambition/7kaaInterface/draw.hh
@@ -64,7 +64,7 @@ void buildingAnimationFrame(
  * @param firm A pointer to the Firm for which to draw the icon.
  */
 void buildingBuilderIcon(
-  Firm* firm
+  const Firm* firm
 );
 
 /**
@@ -73,7 +73,7 @@ void buildingBuilderIcon(
  * @param firm The Firm for which to draw a hitbar.
  */
 void buildingHitBar(
-  Firm* firm
+  const Firm* firm
 );
 
 void buildingLinkLine(
@@ -107,8 +107,15 @@ bool buildingOccupantHitbar(
   const int maximumHitpoints
 );
 
+bool buildingInformationPanel(
+  const Firm* _7kaaFirm
+);
+bool buildingInformationPanel(
+  const Town* _7kaaTown
+);
+
 void buildingProgressBar(
-  Firm* firm
+  const Firm* firm
 );
 
 void buildingProgressBar(
@@ -123,7 +130,7 @@ void buildingRallyButton(
 );
 
 void buildingRallyPoint(
-  Firm* _7kaaFirm
+  const Firm* _7kaaFirm
 );
 void buildingRallyPoint(
   Town* _7kaaTown